feat: settings menu module
