+++ apps/frontend/core-dapp/src/components/SettingsMenu/index.tsx
@@ -1,0 +1,3 @@
+import SettingsDropdown from './SettingsDropdown'
+
+export default SettingsDropdown
